HW13 is completed

+++ hw12_13_14_15_calendar/migrations/00001_add_events_table.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS events (
     end_date TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     description TEXT,
     user_id UUID,
-    notification_time INT
+    notification_time BIGINT
 );
 
 -- +goose Down
